% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainconn.R
\name{brainconn}
\alias{brainconn}
\title{Title}
\usage{
brainconn(atlas, background = "ICBM152", view = "top", conmat = NULL,
  interactive = F, node.size = 4, node.color = "network",
  all.nodes = FALSE, edge.color = "black", edge.alpha = 0.8,
  edge.width = 1, edge.color.weighted = FALSE, labels = FALSE,
  show.legend = TRUE, thr = NULL, uthr = NULL,
  scale.edge.width = NULL, label.size = 1.5, background.alpha = 1,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
