% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_atlas.R
\name{check_atlas}
\alias{check_atlas}
\title{Title}
\usage{
check_atlas(atlas)
}
\arguments{
\item{atlas}{a .data.frame with a 4 collums named: c("ROI.Name", "x.mni", "y.mni", "z.mni"), see \code{vignette("brainconn")}}
}
\value{
a message
}
\description{
Title
}
\examples{
\dontrun{
library(brainconn)
check_atlas(custom_atlas_example)
}

}
