% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggconn.R
\name{ggconn}
\alias{ggconn}
\title{ggconn}
\usage{
ggconn(atlas, background = "ICBM152", view = "top", conmat = NULL,
  interactive = F, node.size = 6, node.color = "network",
  all.nodes = FALSE, edge.color = "black", edge.alpha = 0.8,
  edge.width = 1, labels = FALSE, show.legend = TRUE, thr = NULL,
  uthr = NULL, scale.edge.width = NULL, label.size = 1.5)
}
\arguments{
\item{atlas}{Brain atlas to use. Inbuilt options include "aal91" "schafer300_n7"}

\item{label.size}{}
}
\description{
ggconn
}
\examples{
ggConnectome(atlas ="schafer300_n7", conmat=x, node.color = "network")
\dontrun{
ggConnectome(atlas ="schafer300_n7", conmat=x, node.color = "network")
}
}
